Drop the single-impl IdProviderControllerScript interface
The interface had one implementation and was not an OSGi service contract, so it
was pure indirection. Collapse it: promote IdProviderControllerScriptImpl to the
clean name IdProviderControllerScript (concrete class). The factory
(IdProviderControllerScriptFactory) stays an interface - it is an OSGi service.
